add it as canonicalization
