+++ mlir/include/mlir/Dialect/SCF/Passes.h
@@ -35,9 +35,6 @@ std::unique_ptr<Pass> createParallelLoopSpecializationPass();
 std::unique_ptr<Pass>
 createParallelLoopTilingPass(llvm::ArrayRef<int64_t> tileSize = {});
 
-/// Convert scf.while to scf.for.
-std::unique_ptr<Pass> createWhileOpToForOpPass();
-
 //===----------------------------------------------------------------------===//
 // Registration
 //===----------------------------------------------------------------------===//
