+++ mlir/include/mlir/Dialect/SCF/Passes.td
@@ -44,9 +44,4 @@ def SCFParallelLoopTiling : FunctionPass<"parallel-loop-tiling"> {
   let dependentDialects = ["AffineDialect"];
 }
 
-def SCFWhileToSCFFor : FunctionPass<"scf-while-to-scf-for"> {
-  let summary = "Covert SCF while op to SCF for op";
-  let constructor = "mlir::createWhileOpToForOpPass()";
-}
-
 #endif // MLIR_DIALECT_SCF_PASSES
